REF: move helper functions to tools.linalg
