+++ statsmodels/tools/linalg.py
@@ -75,3 +75,101 @@ def stationary_solve(r, b):
         x = x[:, 0]
 
     return x
+
+
+def transf_constraints(constraints):
+    """use QR to get transformation matrix to impose constraint
+
+    Parameters
+    ----------
+    constraints : ndarray, 2-D
+        restriction matrix with one constraints in rows
+
+    Returns
+    -------
+    transf : ndarray
+        transformation matrix to reparameterize so that constraint is
+        imposed
+
+    Notes
+    -----
+    This is currently and internal helper function for GAM.
+    API not stable and will most likely change.
+
+    The code for this function was taken from patsy spline handling, and
+    corresponds to the reparameterization used by Wood in R's mgcv package.
+
+    See Also
+    --------
+    statsmodels.base._constraints.TransformRestriction : class to impose
+        constraints by reparameterization used by `_fit_constrained`.
+
+    """
+
+    from scipy import linalg
+    # constraints = spsc._constraints
+    m = constraints.shape[0]
+    q, r = linalg.qr(np.transpose(constraints))
+    transf = q[:, m:]
+    return transf
+
+
+def matrix_sqrt(mat, inverse=False, full=False, nullspace=False, threshold=1e-15):
+    """matrix square root for symmetric matrices
+
+    Usage is for decomposing a covariance function S into a square root R
+    such that
+
+        R' R = S if inverse is False, or
+        R' R = pinv(S) if inverse is True
+
+    Parameters
+    ----------
+    mat : array_like, 2-d square
+        symmetric square matrix for which square root or inverse square
+        root is computed.
+        There is no checking for whether the matrix is symmetric.
+        A warning is issued if some singular values are negative, i.e.
+        below the negative of the threshold.
+    inverse : bool
+        If False (default), then the matrix square root is returned.
+        If inverse is True, then the matrix square root of the inverse
+        matrix is returned.
+    full : bool
+        If full is False (default, then the square root has reduce number
+        of rows if the matrix is singular, i.e. has singular values below
+        the threshold.
+    nullspace: bool
+        If nullspace is true, then the matrix square root of the null space
+        of the matrix is returned.
+    threshold : float
+        Singular values below the threshold are dropped.
+
+    Returns
+    -------
+    msqrt : ndarray
+        matrix square root or square root of inverse matrix.
+
+    """
+    # see also scipy.linalg null_space
+    u, s, v = np.linalg.svd(mat)
+    if np.any(s < -threshold):
+        import warnings
+        warnings.warn('some singular values are negative')
+
+    if not nullspace:
+        mask = s > threshold
+        s[s < threshold] = 0
+    else:
+        mask = s < threshold
+        s[s > threshold] = 0
+
+    sqrt_s = np.sqrt(s[mask])
+    if inverse:
+        sqrt_s = 1 / np.sqrt(s[mask])
+
+    if full:
+        b = np.dot(u[:, mask], np.dot(np.diag(sqrt_s), v[mask]))
+    else:
+        b = np.dot(np.diag(sqrt_s), v[mask])
+    return b
